+++ device_info_windows.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"unsafe"
 
-	"github.com/zhuzxdev/gocan/raw"
+	"github.com/zhuzx17/gocan/raw"
 )
 
 func getDeviceInfo(ch Channel) (DeviceInfo, error) {