+++ examples/01_send_classical/main.go
@@ -11,7 +11,7 @@ import (
 	"flag"
 	"log"
 
-	"github.com/zhuzxdev/gocan"
+	"github.com/zhuzx17/gocan"
 )
 
 func main() {