+++ examples/06_multi_channel/main.go
@@ -13,7 +13,7 @@ import (
 	"os/signal"
 	"sync"
 
-	"github.com/zhuzxdev/gocan"
+	"github.com/zhuzx17/gocan"
 )
 
 func main() {