+++ examples/07_filter/main.go
@@ -14,7 +14,7 @@ import (
 	"os/signal"
 	"time"
 
-	"github.com/zhuzxdev/gocan"
+	"github.com/zhuzx17/gocan"
 )
 
 func main() {