+++ examples/10_using_raw/main.go
@@ -13,7 +13,7 @@ import (
 	"log"
 	"unsafe"
 
-	"github.com/zhuzxdev/gocan/raw"
+	"github.com/zhuzx17/gocan/raw"
 )
 
 func main() {