+++ examples/11_busgroup_socketcan/main.go
@@ -13,7 +13,7 @@ import (
 	"os"
 	"os/signal"
 
-	"github.com/zhuzxdev/gocan"
+	"github.com/zhuzx17/gocan"
 )
 
 func main() {